backend/pyproject.toml
[project]
name = "family-health-backend"
version = "0.1.0"
description = "Family health local backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "sqlalchemy>=2.0.30",
  "httpx>=0.27.0",
  "pydantic-settings>=2.4.0",
  "pyjwt>=2.9.0",
  "argon2-cffi>=23.1.0",
  "python-multipart>=0.0.9",
  "openpyxl>=3.1.5",
  "pypdf>=4.3.1",
  "python-pptx>=1.0.2",
  "xlrd>=2.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.0",
  "httpx>=0.27.0",
  "ruff>=0.6.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 100
target-version = "py311"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]
